+++ packages/core/src/crypto/index.js
@@ -1,6 +1,7 @@
 export * from './kdf.js';
 export * from './aead.js';
 export * from './mnemonic.js';
+export * from './counterwallet.js';
 export * from './hd.js';
 export * from './wif.js';
 export * from './walletBlob.js';